Permit nested arrays as computation output (#12)
Symbols and tokens are now objects instead of arrays to avoid conflict with the nested arrays produced by the user. A new calculator function that can run both sync and async is also implemented. Adds version control to the database to admin-panel communication. One readily available result of this is that one can now render lists from data, including lists of lists of elements.